Fix race condition with typeahead find. Because typeahead find is a async operation, the server may return results in milliseconds or seconds, and if the user presses "enter" before the typeahead suggestions appear they could get completely different results than if they press enter immediately after (before they get a chance to even see it), causing substantial confusion.
For example if search for john@company.com and presses enter immediately, if the typeahead suggestions popped up for 1ms (or while the user is looking at their keyboard typing) before they pressed enter with "jdoe@othercompany.com" selected, the user would see results for the 2nd email address which would be completely unexpected to them.

Industry standard practice (ie: searching in web browers, twitter, youtube, etc... ) with typeahead find is it should never default to some other value than what the user is typing, therefore the user should always have to select the suggestion that they want. If they don't select a suggestion, the search they typed in should be used instead.

This required the modification of bootstrap-typeahead.js simply because this is an extremely old bootstrap plugin and no longer supported or recommended to be used. Newer typeahead plugins either default the way we want or have a option to enable/disable selection by default.